Move jats transformers into subfolder.
