+++ src/article/converter/transform/jats/NormalizeContribGroup.js
