+++ src/article/converter/transform/jats/NormalizeFn.js
