+++ src/article/converter/transform/jats/NormalizeRefList.js
