+++ src/article/index.js
@@ -12,5 +12,3 @@ export { default as ArticlePlugin } from './ArticlePlugin'
 export { default as ArticleSerializer } from './ArticleSerializer'
 export { default as ArticleSession } from './ArticleSession'
 export { default as InternalArticleDocument } from './InternalArticleDocument'
-export { default as JATS } from './JATS'
-export { default as TextureJATS } from './TextureJATS'
